Finish model class and update agent class. Test model class
I modified the model class to have it automatically change the size of the fc start based on the screen size. Also, instead of hardcoding the number of actions to 4, it is variable to allow us to change the game without any problems. We added an error check during the forward pass, and we normalize the values to help training.

In agent we create a method to get actions, a method to update epsilon, a method to add memories to the memory buffer, and a method to replay the memories